requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.0
redis==5.0.1
anthropic==0.54.0
httpx==0.25.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
boto3==1.34.0
elevenlabs==0.2.26
youtube-search-python==1.6.6
sse-starlette==1.6.5
aiofiles==23.2.1
jsonschema==4.20.0
pytest==7.4.3
pytest-asyncio==0.21.1 